requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pymysql
python-multipart
python-jose[cryptography]